requirements.txt
autogen-agentchat~=0.2
groq
flaml[automl]